Added an evaluator, exceptions, parser and tests.
Using the nodes.Visitor, I've created the Evaluator.
Created a parser, takes in tokens from lexer and outputs an AST from the
nodes in node module.
Wrote a few tests, doesn't cover all cases, this will be improved in
later and created a base for my exceptions in all of this project.