Added Compiler, instructions and a VirtualMachine.
With a few instructions, I have made the VM and the compiler.
They operate on bytecode of a specific format, that is, it has to
sections, the constants section and program section. Both sections end
with a special instruction, EOS (end of section) to mark the end of one
and the beginning of the other. The compiler takes in AST and outputs
bytecode while the VM takes in bytecode and outputs the last value in
the operational stack unlike the evaluator which simply walked the AST
executing each node as it goes.