Added a facade and an astprinter.
The execution chain is quite long and tedious to create all the time:
create a lexer, then a parser then an evaluator then pipeline the expr
through the lexer, parser and finally to the evaluator just to get a
an answer. I have fixed this by creating a Facade class, ExprEvaluator,
to hide away the execution pipeline by exposing 3 methods necessary to
execute an expression from start to end: compile, exec and eval.

The compile method compiles an expression to bytecode, the exec method
executes bytecode while eval executes the passed expression via the
TreeWalkInterpreter, Evaluator. Other objects like then Compiler,
Evaluator, VirtualMachine, Lexer and Parser are properties of the Facade
class which initializes them as needed so you don't have to.

A basic, but pretty useful ASTPrinter has been created and added to
the printer module alongside ByteCodePrinter. It simply prints the AST
produced by the Parser.
Example:
Expression: 100 / 20 * 4 + 50 - 2 ^ 10 / 128 * 7

AST Tree:
-
+           *
*        50 /        7
/      4    ^    128
100 20      2 10

If read carefully, it represents the above expression.