Fixed and improved the project quality and readability
From my experience in coding large projects, readability and a
comprehensive interfaces are rulers of the land, if the code isn't
readable, good luck finding, patching and maintaining the project. So,
I've worked hard along with mypy to find and fix this projects in terms
of readability by employing better names, maintainability by ensuring
correct type hints and lastly, made operations simple and clear by
removing unncessary steps to get the job done. For the first time mypy
has really helped in this area, though I do not understand the tool as
much but I know it is a valuable tool to own.